backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.6.0
python-multipart==0.0.6
python-dotenv
langchain
langchain-core
langchain-groq
langchain-google-genai
httpx
sqlalchemy
langchain-core
langchain-google-genai
langchain
python-dotenv
httpx
psycopg2-binary
